+++ src/OpenTelemetry.Instrumentation.SqlClient/CHANGELOG.md
@@ -11,6 +11,20 @@
   `db.query.summary` length limit is reached.
   ([#4929](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4929))
 
+* Updated Semantic Conventions version to
+  [v1.44.0](https://github.com/open-telemetry/semantic-conventions/blob/v1.44.0/docs/db/sql-server.md).
+  ([#4625](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4625))
+
+* Add the `db.response.returned_rows` attribute to query spans when the
+  `OTEL_DOTNET_EXPERIMENTAL_SQLCLIENT_ENABLE_RECORD_RETURNED_ROWS` environment
+  variable is set to `true`. The value is derived from the SqlClient connection
+  statistics. Not supported on .NET Framework.
+  ([#4625](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4625))
+
+* Fixed the `db.client.operation.duration` metric reporting an incorrect duration
+  for commands which are not recorded as spans.
+  ([#4625](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4625))
+
 ## 1.17.0
 
 Released 2026-Jul-17